Hope fully the main menu and volume change in VR scene
I added a main menu that can quit the game start the game and control volume. all in the Main menu scene. I change the volume connection in the VR scene to the master volume I made in the Main menu I hope when I merge this nothing gets erased.