I added a tutorial video ish which I want to discuss
I added a tutorial video ish yes you can it the tutorial button and hit play and pause and quit. but we should'nt have a tutorial for a game thats not finished and wont look like this in the first place. Good news it I can always switch out the video so at least its ready to go.