MainMenu1 Remixed to work properly in VR
Added the Mainmenu 1 which is whats used for a vr menu. and added scene of level 1 / Library. If the play button does not work make sure to go to project settings and put mainmenu 1 in the top added scenes where it has a ZERO to the right of it then add the Game scene which houses the word saber. make sure it has a 1 to the right of it then back out of the menu and try again and it should work.